pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["DSulzhits <dsulzhits@gmail.com>"]
readme = "README.md"
packages = [{ include = "habit_tracker" }]

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
django = "^4.2.4"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.7"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.0"
django-cors-headers = "^4.2.0"
requests = "^2.31.0"
redis = "^5.0.0"
celery = "^5.3.1"
django-celery-beat = "^2.5.0"
telebot = "^0.0.5"
python-telegram-bot = "^20.4"
django-crontab = "^0.7.1"
drf-yasg = "^1.21.7"


[tool.poetry.group.other.dependencies]
ipython = "^8.14.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
